tests/submission/W3C/domhtml/HTMLAppletElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Applet</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-applet-align" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<applet align="bottom" alt="Applet Number 1" archive="" code="org/w3c/domts/DOMTSApplet.class" codebase="applets" height="306" hspace="0" name="applet1" vspace="0" width="301"></applet>
</p>


<script>
var domtest = async_test("The align, alt, archive, code, height, hspace, name, object, vspace, and width IDL attributes must reflect the respective content attributes of the same name. For the purposes of reflection, the applet element's object content attribute is defined as containing a URL.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valign;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("applet");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      valign = testNode.align;

      assert_equals(valign.toLowerCase(),"bottom".toLowerCase());
       
    domtest.done();
}));

</script>
<p>The align, alt, archive, code, height, hspace, name, object, vspace, and width IDL attributes must reflect the respective content attributes of the same name. For the purposes of reflection, the applet element's object content attribute is defined as containing a URL.</p>
<div id="log"></div>
</body></html>
